+++ crates/bevy_app/src/app.rs
@@ -1,4 +1,5 @@
 use crate::app_builder::AppBuilder;
+use crate::DefaultTaskPoolOptions;
 use bevy_ecs::{ParallelExecutor, Resources, Schedule, World};
 
 #[allow(clippy::needless_doctest_main)]
@@ -63,7 +64,11 @@ impl App {
     }
 
     pub fn run(mut self) {
-        ParallelExecutor::initialize_pools(&mut self.resources);
+        // Setup the default bevy task pools
+        self.resources
+            .get_cloned::<DefaultTaskPoolOptions>()
+            .unwrap_or_else(DefaultTaskPoolOptions::default)
+            .create_default_pools(&mut self.resources);
 
         self.startup_schedule.initialize(&mut self.resources);
         self.startup_executor.run(
