+++ src/lib.rs
@@ -53,12 +53,12 @@ pub use bevy_property as property;
 pub use bevy_render as render;
 pub use bevy_scene as scene;
 pub use bevy_sprite as sprite;
+pub use bevy_tasks as tasks;
 pub use bevy_text as text;
 pub use bevy_transform as transform;
 pub use bevy_type_registry as type_registry;
 pub use bevy_ui as ui;
 pub use bevy_window as window;
-pub use bevy_tasks as tasks;
 
 #[cfg(feature = "bevy_audio")]
 pub use bevy_audio as audio;
